T-ESP-900-28649-VirtuHealth-4/IA_BD/api/requirements.txt
asgiref==3.6.0
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.2
django-cors-headers==3.14.0
djangorestframework==3.14.0
drf-spectacular==0.26.1
idna==3.4
inflection==0.5.1
jsonschema==4.17.3
pyrsistent==0.19.3
pytz==2023.3
PyYAML==6.0
requests==2.28.2
sqlparse==0.4.3
toml==0.10.2
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.15
